+++ IntoTheSpireverseCode/Cards/ShadowIronclad/Titanic.cs
@@ -21,7 +21,7 @@ public sealed class Titanic() : ShadowIroncladCard(3, CardType.Power, CardRarity
     protected override async Task OnPlay(PlayerChoiceContext choiceContext, CardPlay cardPlay)
     {
         await CreatureCmd.TriggerAnim(Owner.Creature, "Cast", Owner.Character.CastAnimDelay);
-        await CreatureCmd.GainMaxHp(Owner.Creature, (decimal)DynamicVars.MaxHp.IntValue);
+        await CreatureCmd.GainMaxHp(Owner.Creature, DynamicVars.MaxHp.IntValue);
     }
 
     protected override void OnUpgrade() => DynamicVars.MaxHp.UpgradeValueBy(1m);
